[sgh-12-30] Changed product rent price type from Integer to Double

+++ src/main/java/edu/codespring/sportgh/dto/ProductInDTO.java
@@ -17,7 +17,7 @@ public class ProductInDTO {
     private String description;
 
     @NotNull
-    private Integer rentPrice;
+    private Double rentPrice;
 
     @NotNull
     private Long subCategoryId;
